+++ modules/organization/main.tf
@@ -53,7 +53,7 @@ locals {
     type => {
       for name, sink in local.logging_sinks :
       name => sink
-      if sink.grant && sink.type == type
+      if sink.iam && sink.type == type
     }
   }
 }
@@ -221,9 +221,10 @@ resource "google_logging_organization_sink" "sink" {
   for_each = local.logging_sinks
   name     = each.key
   #description = "${each.key} (Terraform-managed)"
-  org_id      = local.organization_id_numeric
-  destination = "${local.sink_type_destination[each.value.type]}/${each.value.destination}"
-  filter      = each.value.filter
+  org_id           = local.organization_id_numeric
+  destination      = "${local.sink_type_destination[each.value.type]}/${each.value.destination}"
+  filter           = each.value.filter
+  include_children = each.value.include_children
 }
 
 resource "google_storage_bucket_iam_binding" "gcs-sinks-binding" {
@@ -248,13 +249,6 @@ resource "google_pubsub_topic_iam_binding" "pubsub-sinks-binding" {
   role     = "roles/pubsub.publisher"
   members  = [google_logging_organization_sink.sink[each.key].writer_identity]
 }
-
-# resource "google_storage_bucket_iam_binding" "gcs-sinks-bindings" {
-#   for_each = local.sink_grants["gcs"]
-#   bucket   = each.value.destination
-#   role     = "roles/storage.objectCreator"
-#   members  = [google_logging_organization_sink.sink[each.key].writer_identity]
-# }
 
 resource "google_logging_organization_exclusion" "logging-exclusion" {
   for_each    = coalesce(var.logging_exclusions, {})