+++ modules/organization/variables.tf
@@ -102,10 +102,11 @@ variable "firewall_policy_attachments" {
 variable "logging_sinks" {
   description = "Logging sinks to create for this organization."
   type = map(object({
-    destination = string
-    type        = string
-    filter      = string
-    grant       = bool
+    destination      = string
+    type             = string
+    filter           = string
+    iam              = bool
+    include_children = bool
   }))
   default = {}
 }