+++ tests/modules/project/fixture/variables.tf
@@ -93,3 +93,18 @@ variable "services" {
   type    = list(string)
   default = []
 }
+
+variable "logging_sinks" {
+  type = map(object({
+    destination = string
+    type        = string
+    filter      = string
+    iam         = bool
+  }))
+  default = {}
+}
+
+variable "logging_exclusions" {
+  type    = map(string)
+  default = {}
+}